+++ CHANGELOG.md
@@ -11,8 +11,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ### Added
 
-- Add `--sort` option to `lighthouse:print-schema` command to sort the compiled
-  schema before printing https://github.com/nuwave/lighthouse/pull/2696
+- Add option `--sort` to command `lighthouse:print-schema` https://github.com/nuwave/lighthouse/pull/2696
 
 ## v6.60.0