+++ CHANGELOG.md
@@ -9,6 +9,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 ## Unreleased
 
+## v6.61.0
+
 ### Added
 
 - Add option `--sort` to command `lighthouse:print-schema` https://github.com/nuwave/lighthouse/pull/2696